week_03/day_02/pair_programming/week3_day1-main/src/main/java/Midfielder.java
public class Midfielder extends Player implements Shooters, Tackles {

    public Midfielder(String name, int age, int wage, int goalsScored, double shootingAccuracy) {
        super(name, age, wage, goalsScored, shootingAccuracy);
    }

    public void shoot() {
        if(Math.random() < shootingAccuracy) {
            this.goalsScored++;
        }
    }

    public void shoot(boolean isPenalty) {
        if((Math.random() * 2 ) < shootingAccuracy) {
            this.goalsScored++;
        }
    }

    @Override
    public String tackle() {
        return "Standing up tackle.";
    }

}
